refactor(plugin): 实例生命周期端点独立成篇并收束安装用例规模
plugin.py 随多版本端点累积到 1244 行，越过文件 1000 行预算。分身创建与恢复、
启停、彻底清理与卸载本就是同一组关注点，独立为 plugininstance.py 后按
pluginfolder 的既有先例并入同一路由器，路径、方法与权限依赖逐项不变。

install.py 的 PluginInstallCommand 与 __execute_locked 本就压在预算线上，本次
新增的装载登记把两者顶了出去：重载运行态与刷新注册合成一个私有方法，载荷身份
比较中不碰磁盘的部分与 __await_side_effect 上移模块级，两项指标回到线下。
